Separate out scenario validation into two functions.
This commit separates the request and validation logic into two
functions so that a TS_OM_VALIDATE_URL is not required in order to
validate scenarios. This allows for creating and running tests without
having to work around the fact that a server is not running and hosting
the om_validate app at a specific host:port endpoint.